Move classes to package
