+++ src/dogs/src/test/java/org/cloudfoundry/identity/acceptance/SamlLoginAcceptanceTests.java
@@ -1,4 +1,4 @@
-/*******************************************************************************
+package org.cloudfoundry.identity.acceptance; /*******************************************************************************
  *     Cloud Foundry
  *     Copyright (c) [2009-2017] Pivotal Software, Inc. All Rights Reserved.
  *
@@ -12,6 +12,8 @@
  *******************************************************************************/
 import org.hamcrest.Matchers;
 import org.junit.After;
+import org.junit.Assert;
+import org.junit.Assume;
 import org.junit.Before;
 import org.junit.Test;
 import org.junit.runner.RunWith;
@@ -21,13 +23,13 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.test.context.ContextConfiguration;
 import org.springframework.test.context.junit4.SpringJUnit4ClassRunner;
+
+import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertThat;
 
 @RunWith(SpringJUnit4ClassRunner.class)
 @ContextConfiguration(classes = DefaultAcceptanceTestConfig.class)
-public class LdapLoginAcceptanceTests {
-
-
+public class SamlLoginAcceptanceTests {
     @Value("${BASE_URL}")
     private String baseUrl;
 
@@ -36,23 +38,32 @@ public class LdapLoginAcceptanceTests {
 
     @Autowired
     private WebDriver webDriver;
+    private String url;
 
     @Before
     @After
     public void clearWebDriverOfCookies() throws Exception {
+        url = protocol + baseUrl;
         webDriver.get(protocol + baseUrl + "/logout.do");
+        webDriver.manage().deleteAllCookies();
     }
 
     @Test
-    public void ldap_login_IsSuccessful() {
-        webDriver.findElement(By.name("username")).sendKeys("marissa-ldap");
-        webDriver.findElement(By.name("password")).sendKeys("marissa-ldap");
-        webDriver.findElement(By.xpath("//input[@value='Sign in']")).click();
+    public void testUrlSamlPhpLoginAWS() throws Exception {
+        Assume.assumeTrue("This test is against AWS environment", url.contains(".identity.cf-app.com"));
+        samlPhpLogin("Log in with Simple SAML PHP URL");
+    }
 
+    private void samlPhpLogin(String linkText) {
+        webDriver.get(protocol + baseUrl + "/login");
+        Assert.assertEquals("Cloud Foundry", webDriver.getTitle());
+        webDriver.findElement(By.xpath("//a[text()='" + linkText + "']")).click();
+        webDriver.findElement(By.xpath("//h2[text()='Enter your username and password']"));
+        webDriver.findElement(By.name("username")).clear();
+        webDriver.findElement(By.name("username")).sendKeys("marissa");
+        webDriver.findElement(By.name("password")).sendKeys("koala");
+        webDriver.findElement(By.xpath("//input[@value='Login']")).click();
+        assertEquals("marissa@test.org", webDriver.findElement(By.xpath("//div[@class='dropdown-trigger']")).getText());
         assertThat(webDriver.findElement(By.cssSelector("h1")).getText(), Matchers.containsString("Where to?"));
-
-        webDriver.get(protocol + baseUrl + "/logout.do");
-
-        assertThat(webDriver.findElement(By.cssSelector("h1")).getText(), Matchers.containsString("Welcome!"));
     }
 }
