+++ src/dogs/src/test/java/org/cloudfoundry/identity/acceptance/SiteMinderIntegrationTests.java
@@ -1,3 +1,5 @@
+package org.cloudfoundry.identity.acceptance;
+
 import org.hamcrest.Matchers;
 import org.junit.*;
 import org.junit.runner.RunWith;
